backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.8.2
pydantic-settings==2.4.0
textblob==0.18.0.post0
pytest==8.3.2
httpx==0.27.0
